requirements.txt
certifi==2019.3.9
chardet==3.0.4
cloudpickle==0.8.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dask==1.1.4
decorator==4.4.0
Django==1.11.9
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.2
numpy==1.14.1
openapi-codec==1.3.2
opencv-python==4.0.0.21
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.3
PyWavelets==1.0.2
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
toolz==0.9.0
uritemplate==3.0.0
urllib3==1.24.1
